refactor(task_executor): move test module to task_executor/tests.rs
Extracts the 341-line #[cfg(test)] mod tests block from task_executor.rs
into task_executor/tests.rs. The file declares the tests as a child module
via #[cfg(test)] mod tests; so super::* still resolves correctly.
